+++ NEWS.md
@@ -3,6 +3,7 @@
 - Implement orbit number and relative orbit number to queries
 - Implement max records to queries
 - Implement reading of bands from zip files
+- Implement no check option for creation of Tiles
 
 
 # theiaR 0.1.2